+++ src/Products/PloneMeeting/migrations/migrate_to_4215.py
@@ -86,7 +86,7 @@ class Migrate_To_4215(Migrator):
                 cfg.registerPortalTypes()
         logger.info('Done.')
 
-    def _fixWSCConfigAndCleanBrokenAnnexes(self):
+    def _fixWSCConfig(self):
         """WSC could sometimes break quickupload leading to broken annexes.
            Adapt WSC config so it is no more used in quickupload and clean broken annexes."""
         logger.info("Updating WSC config and removing broken annexes...")
@@ -95,8 +95,6 @@ class Migrate_To_4215(Migrator):
             set_disable_autosearch_in(
                 u'["#form-widgets-title", "#form-widgets-description", '
                 u'".select2-focusser", ".select2-input"]')
-            # remove broken annexes
-            self._removeBrokenAnnexes()
         logger.info('Done.')
 
     def _updateWFWriteMarginalNotesPermission(self):
@@ -118,6 +116,8 @@ class Migrate_To_4215(Migrator):
     def run(self, extra_omitted=[], from_migration_to_4200=False):
 
         logger.info('Migrating to PloneMeeting 4215...')
+        # remove broken annexes before upgrading collective.iconifiedcategory
+        self._removeBrokenAnnexes()
         if not from_migration_to_4200:
             # this will upgrade collective.contact.core especially
             # that reinstalls imio.fpaudit, that will itself reinstall collective.documentgenerator
@@ -130,9 +130,9 @@ class Migrate_To_4215(Migrator):
         self._initGroupsManagingItemToCfgItemWFValidationLevels()
         self._migrateItemsWaitingAdviceState()
         self._updateSearchItemsOfMyCommitteesSearchesCondition()
+        self._fixWSCConfig()
         self._updateConfigCustomAdvisersDataGrid()
         self._reloadMeetingConfigsForItemWorkflows()
-        self._fixWSCConfigAndCleanBrokenAnnexes()
         self._updateWFWriteMarginalNotesPermission()
         logger.info('Migrating to PloneMeeting 4215... Done.')